% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{adjust_matches}
\alias{adjust_matches}
\title{Adjust Tables for already matched names}
\usage{
adjust_matches(.tab_matches, .tab)
}
\arguments{
\item{.tab_matches}{A Dataframe generated by match_name() or match_substring()}

\item{.tab}{A Dataframe generated by extract_legal_form()}
}
\value{
A Dataframe
}
\description{
Adjust Tables for already matched names
}
