% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OTHER_FUNCTIONS.R
\name{check_legal_form}
\alias{check_legal_form}
\title{Check if Entities have the same legal form}
\usage{
check_legal_form(.tab, .tab_lf)
}
\arguments{
\item{.tab}{A dataframe generated by match_names()}

\item{.tab_lf}{A dataframe generated by make_legal_form_table()}
}
\value{
A dataframe with one additional column: same_legal_form
}
\description{
Check if Entities have the same legal form
}
\examples{

.path <- system.file("extdata", "test_tables.xlsx", package = "RFirmMatch")
.tab1 <- openxlsx::read.xlsx(.path, 1)
.tab1 <- .tab1 \%>\% prepare_tables() \%>\% extract_legal_form(make_legal_form_table())

.tab2 <- openxlsx::read.xlsx(.path, 2)
.tab2 <- .tab2 \%>\% prepare_tables() \%>\% extract_legal_form(make_legal_form_table())

.tab <- match_name(.tab1, .tab2, name_clean, "full")
.tab_lf <- make_legal_form_table()

check_legal_form(.tab, .tab_lf)

rm(.tab, .tab_lf)

}
