% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{extract_legal_form}
\alias{extract_legal_form}
\title{Extract Legal Forms from Company Names}
\usage{
extract_legal_form(.table, .table_le = NULL, .progress = FALSE)
}
\arguments{
\item{.table}{A table prepared by prepare_tables()}

\item{.table_le}{Additional Legal Entity Table}

\item{.progress}{Show progress bar?}
}
\value{
A Dataframe with 10 columns: \cr
Column 1: 'table_id' \<int\>: ... \cr
Column 2: 'id' \<chr\>: A column with the ID(s) \cr
Column 3: 'name' \<chr\>: A column with standardized names \cr
Column 4: 'name_adj' \<chr\>: A column with firm names w/o legal forms \cr
Column 5: 'country_code' \<chr\>: A column with ISO 3166_2 Alpha 3 country codes \cr
Column 6: 'continent_code' \<chr\>: A column with continent codes \cr
Column 7: 'n_id' \<int\>: A column with the number of matched IDs to the firm name \cr
Column 8: 'legal_form' \<chr\>: A column with extracted legal form \cr
Column 9: 'legal_form_id' \<list\>: A column with legal form IDs \cr
column 10: country_code_legal_form' \<list\>: A column with corresponding ISO 3166_2 Alpha 3 country codes to the legal forms
}
\description{
Extract Legal Forms from Company Names
}
\examples{
table <- tibble::tibble(
 id = 1:2,
 country_code = c("DEU", NA),
  name1 = c("BASF AG", "BASF SE"),
  name2 = c("BASF Germany AG", "BASF Germany SE")
)
tab1 <- prepare_tables(table, "firm")
tab2 <- prepare_tables(table, "match")

extract_legal_form(tab1)
extract_legal_form(tab2)
}
