% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions_2.R
\name{extract_legal_form2}
\alias{extract_legal_form2}
\title{Extract Legal Forms from Company Names}
\usage{
extract_legal_form2(.tab, .tab_lf = NULL, .workers = 1)
}
\arguments{
\item{.tab}{A table prepared by prepare_tables()}

\item{.tab_lf}{Additional Legal Form Table}

\item{.workers}{Number of parallel workers}
}
\value{
A datframe
}
\description{
Extract Legal Forms from Company Names
}
\examples{
.tab <- table <- prepare_tables2(RFirmMatch::table_test0)

## DEBUG
.tab_lf = NULL
.workers = 1


extract_legal_form2(table)
}
