% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{helper_get_substrings_list}
\alias{helper_get_substrings_list}
\title{Helper Function: Cut Strings}
\usage{
helper_get_substrings_list(.string, .min_char)
}
\arguments{
\item{.string}{A charcter atring}

\item{.min_char}{Number between 0 and 1 exclusively (Determines the minimum percent of the string to cut)}
}
\value{
A list with cutted strings
}
\description{
Helper Function: Cut Strings
}
\examples{
helper_get_substrings_list("TEST STRING", .8)
}
