% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{helper_get_substrings_table}
\alias{helper_get_substrings_table}
\title{Helper Function: Cut String}
\usage{
helper_get_substrings_table(.tab, .col_match, .min_char)
}
\arguments{
\item{.tab}{A table}

\item{.col_match}{The column with the string to be cut}

\item{.min_char}{Number between 0 and 1 exclusively (Determines the minimum percent of the string to cut)}
}
\value{
A Dataframe
}
\description{
Helper Function: Cut String
}
\examples{

tab <- tibble::tibble(id = 1, name = "BADISCHE ANELIN UND SODAFABRIK")
helper_get_substrings_table(tab, name, .5)
rm(tab)
}
