% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{helper_standardize_names}
\alias{helper_standardize_names}
\title{Helper Function: Standardize Names}
\usage{
helper_standardize_names(.name)
}
\arguments{
\item{.name}{A character string}
}
\value{
A character string
}
\description{
Helper Function: Standardize Names
}
\examples{
helper_standardize_names("Äzz--jkl\%&\%")
}
