% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OTHER_FUNCTIONS.R
\name{make_legal_form_table}
\alias{make_legal_form_table}
\title{MAke Legal Form Table}
\usage{
make_legal_form_table(.tab_lf = NULL)
}
\arguments{
\item{.tab_lf}{An additional table for legal forms, needs to have at least 3 columns:\cr
legal_form_id: A non-unique identifier grouping different spellings of legal forms\cr
legal_form: A legal form\cr
country_code: country code of the legal form in ISO 3166 Alpha 3 format\cr

If NULL, internal table will be returned}
}
\value{
A dataframe with legal forms
}
\description{
MAke Legal Form Table
}
\examples{

make_legal_form_table()
}
