% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{match_name}
\alias{match_name}
\title{Match company names}
\usage{
match_name(
  .tab0,
  .tab1,
  .col_match,
  .type = c("full", "sub", "approx"),
  .min_char = 0.25,
  .max_dist = 0.1,
  .method = c("osa", "lv", "dl", "hamming", "lcs", "qgram", "cosine", "jaccard", "jw",
    "soundex"),
  .workers = 1,
  .progress = FALSE
)
}
\arguments{
\item{.tab0}{Company Table}

\item{.tab1}{Matching Table}

\item{.col_match}{Column name used for matching}

\item{.type}{c("full", "sub", "approx")}

\item{.min_char}{only used if .type == "sub"}

\item{.max_dist}{only used if .type == "approx"}

\item{.method}{only used if .type == "approx"}

\item{.workers}{Number of parallel workers (only used for .type == "sub")}

\item{.progress}{Show progress bar?}
}
\value{
A Dataframe
}
\description{
Match company names
}
\examples{
.path <- system.file("extdata", "test_tables.xlsx", package = "RFirmMatch")
.tab1 <- openxlsx::read.xlsx(.path, 1)
.tab1 <- .tab1 \%>\% prepare_tables() \%>\% extract_legal_form(make_legal_form_table())

.tab2 <- openxlsx::read.xlsx(.path, 2)
.tab2 <- .tab2 \%>\% prepare_tables() \%>\% extract_legal_form(make_legal_form_table())

match_name(.tab1, .tab2, name_clean, "full")
match_name(.tab1, .tab2, name_clean, "sub")
match_name(.tab1, .tab2, name_clean, "approx")


## DEBUG
.col_match <- quote(name_clean)
.min_char = 0.8
.max_dist = .1
.method = "osa"
.workers = 1
}
