% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions_2.R
\name{match_name2}
\alias{match_name2}
\title{Match company names}
\usage{
match_name2(
  .tab0,
  .tab1,
  .col_match,
  .type = c("full", "sub", "approx"),
  .min_char = 0.25,
  .max_dist = 2,
  .method = c("osa", "lv", "dl", "hamming", "lcs", "qgram", "cosine", "jaccard", "jw",
    "soundex"),
  .workers = 1
)
}
\arguments{
\item{.tab0}{Company Table}

\item{.tab1}{Matching Table}

\item{.col_match}{Column name used for matching}

\item{.type}{c("full", "sub", "approx")}

\item{.min_char}{only used if .type == "sub"}

\item{.max_dist}{only used if .type == "approx"}

\item{.method}{only used if .type == "approx"}

\item{.workers}{Number of parallel workers (only used for .type == "sub")}
}
\value{
A Dataframe
}
\description{
Match company names
}
\examples{
.tab0 <- table0 <-  RFirmMatch::table_test0 \%>\%
prepare_tables2() \%>\%
extract_legal_form2()

.tab1 <- table1 <-  RFirmMatch::table_test1 \%>\%
prepare_tables2() \%>\%
extract_legal_form2()

## DEBUG
.col_match <- quote(name_clean)
.min_char = 0.25
.max_dist = .25
.method = "osa"
.workers = 1
}
