% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions_2.R
\name{prepare_tables2}
\alias{prepare_tables2}
\title{Prepare Firm Name Table}
\usage{
prepare_tables2(.tab, .regex_names = "name")
}
\arguments{
\item{.tab}{A Dataframe with at least 3 columns: \cr
id: A company identifier \cr
...: any number of columns with company names \cr
country: The country of incorporation of the firm}

\item{.regex_names}{A regular expression that identifies the columns for the company names}
}
\value{
A dataframe
}
\description{
Prepare Firm Name Table
}
\examples{
table<- tibble::tibble(
id = 1:3,
country = c("DEU", NA, "DEU"),
name1 = c("BASF AG", "BASF SE", "BASF AG."),
name2 = c("BASF Germany AG", "BASF Germany SE", "XXX")
)

prepare_tables2(table)


}
