% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{standardize_names}
\alias{standardize_names}
\title{Standrdize Firm Names (Helper Function)}
\usage{
standardize_names(.name)
}
\arguments{
\item{.name}{A character string}
}
\value{
A character string
}
\description{
Standrdize Firm Names (Helper Function)
}
\examples{
 standardize_names("BASF Östereich SE.")
}
