% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{table_country}
\alias{table_country}
\title{Country Code Table}
\format{
A Dataframe with 2,305 rows and 8 variables:
\describe{
\item{iso3}{...}
\item{iso2}{...}
\item{iso_name}{...}
\item{name}{...}
\item{type}{...}
\item{source}{...}
\item{continent_code}{...}
\item{continent_name}{...}
}
}
\usage{
table_country
}
\description{
A Dataframe containing Country Information
}
\keyword{datasets}
