% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{table_legal_forms}
\alias{table_legal_forms}
\title{Legal Form Table}
\format{
A Dataframe with 9,740 rows and 4 variables:
\describe{
\item{legal_form_id}{An ID identifying legal forms}
\item{legal_form}{full or abbreviated legal form}
\item{country_code}{ISO 3166_2 Alpha 3 country code for the legal form}
\item{source}{The source of the legal form}
}
}
\usage{
table_legal_forms
}
\description{
A Dataframe containing Legal Forms
}
\keyword{datasets}
