% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{extract_legal_form}
\alias{extract_legal_form}
\title{Extract Legal Forms from Company Names}
\usage{
extract_legal_form(.tab, .tab_lf, .workers = 1, .progress = FALSE)
}
\arguments{
\item{.tab}{A table prepared by prepare_tables()}

\item{.tab_lf}{A dataframe generated by make_legal_form_table()}

\item{.workers}{Number of parallel workers}

\item{.progress}{Show progress bar?}
}
\value{
A datframe
}
\description{
Extract Legal Forms from Company Names
}
\examples{
.path <- system.file("extdata", "test_tables.xlsx", package = "RFirmMatch")
.tab <- openxlsx::read.xlsx(.path)
.tab <- prepare_tables(.tab)
extract_legal_form(.tab, make_legal_form_table())
rm(.tab)


## DEBUG
.tab_lf = NULL
.workers = 2
.progress = FALSE

}
