% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{prepare_tables}
\alias{prepare_tables}
\title{Prepare Firm Name Table}
\usage{
prepare_tables(.tab, .regex_names = "name")
}
\arguments{
\item{.tab}{A Dataframe with at least 3 columns: \cr
id: A company identifier \cr
...: any number of columns with company names \cr
country: The country of incorporation of the firm}

\item{.regex_names}{A regular expression that identifies the columns for the company names}
}
\value{
A dataframe
}
\description{
Prepare Firm Name Table
}
\examples{
.path <- system.file("extdata", "test_tables.xlsx", package = "RFirmMatch")
.tab <- openxlsx::read.xlsx(.path)
prepare_tables(.tab)
rm(.tab)
}
